requirements.txt
streamlit
python-dotenv
google-generativeai
faiss-cpu
numpy
